requirements.txt
openai>=1.30.0
azure-identity
azure-devops
msrest
pandas
requests
python-dotenv
fastapi
uvicorn
pydantic>=2.0
